添加了图片及视频压缩功能， 修改了图片压缩的方式， 添加了点击添加按钮时弹出底部选择框的效果

+++ lib/widgets/camera/widget/take_video.dart
@@ -3,6 +3,7 @@ import 'dart:io';
 import 'package:camerawesome/camerawesome_plugin.dart';
 import 'package:flutter/material.dart';
 import 'package:wechat_camera_picker/wechat_camera_picker.dart';
+import 'package:wechat_moments/utils/compress.dart';
 import 'package:wechat_moments/widgets/camera/widget/count_down.dart';
 
 // ignore: must_be_immutable
@@ -11,10 +12,12 @@ class TakeVideoPage extends StatefulWidget {
 
   Duration? maxVideoDuration;
 
-  TakeVideoPage({Key? key, required this.cameraState, Duration? maxVideoDuration})
+  TakeVideoPage(
+      {Key? key, required this.cameraState, Duration? maxVideoDuration})
       : super(key: key) {
     this.maxVideoDuration = maxVideoDuration ?? _defaultMaxVideoDuration;
   }
+
   final CameraState cameraState;
 
   //
@@ -32,20 +35,29 @@ class _TakeVideoPageState extends State<TakeVideoPage> {
     widget.cameraState.captureState$.listen((event) async {
       if (event != null && event.status == MediaCaptureStatus.success) {
         String filePath = event.filePath;
-        String fileTitle = filePath
-            .split("/")
-            .last;
+        String fileTitle = filePath.split("/").last;
 
         File file = File(filePath);
 
+        // 压缩视频
+        CompressMediaFile getCompressMediaFile =await DuCompress.video(file);
+
+        // 获取压缩后的视频文件
+        File? getVideoFile = getCompressMediaFile.video?.file;
+
+        if(getVideoFile==null){
+          throw "VideoFile error";
+        }
+
         // 转换为AssetEntity
         AssetEntity? asset = await PhotoManager.editor.saveVideo(
-          file,
+          getVideoFile,
           title: fileTitle,
         );
 
         // 删除临时文件
         await file.delete();
+        await getVideoFile.delete();
 
         Navigator.pop(
           context,
@@ -57,12 +69,17 @@ class _TakeVideoPageState extends State<TakeVideoPage> {
 
   Widget _rightArea() {
     //如果cameraState处于正在录制的状态即cameraState是VideoRecordingCameraState，那么才会出现倒计时的组件
-    if (widget.cameraState is VideoRecordingCameraState && widget.maxVideoDuration != null) {
-      return Countdown(time: widget.maxVideoDuration, callback: () {
-        (widget.cameraState as VideoRecordingCameraState).stopRecording();
-      });
+    if (widget.cameraState is VideoRecordingCameraState &&
+        widget.maxVideoDuration != null) {
+      return Countdown(
+          time: widget.maxVideoDuration,
+          callback: () {
+            (widget.cameraState as VideoRecordingCameraState).stopRecording();
+          });
     } else {
-      return const SizedBox(width: 32 + 20 * 2,);
+      return const SizedBox(
+        width: 32 + 20 * 2,
+      );
     }
   }
